tests: Rename integration_test_utils.py to helper_functions.py
The files utils_integration_test.py and integration_test_utils.py had very similar names and the second one did not contain tests.
